Locking object in place
Placing three fingers on the screen locks the object in place. This involved getting rid of the option to remove the whiteboard however. A better solution will be found. This allows for drawing during the demo, but eventually the functionality will change and the remove option can be added back.